+++ src/helper.rs
@@ -1,5 +1,3 @@
-extern crate reqwest;
-
 use regex::Regex;
 use crate::network::Network;
 
@@ -48,23 +46,6 @@ impl UpgradeHelper {
             println!("Invalid target version for {}: {}", self.network, self.target_version); // TODO: return error here instead
             false 
         }
-    }
-
-}
-
-/// Checks if the release for the target version already exists by 
-/// sending a HTTP request to the GitHub release page.
-pub fn check_release_exists(version: String) -> Result<bool, reqwest::Error> {
-    let release_url = format!("https://github.com/evmos/evmos/releases/tag/{}", version);
-
-    let resp = reqwest::blocking::get(release_url)?;
-    match resp.status() {
-        reqwest::StatusCode::NOT_FOUND => {
-            Ok(false)
-        },
-        _ => {
-            Ok(true)
-        },
     }
 }
 
@@ -125,15 +106,5 @@ mod tests {
             "v14.0.0-rc1".to_string(),
         );
         assert_eq!(helper.check_target_version(), false);
-    }
-
-    #[test]
-    fn test_check_release_exists_pass() {
-        assert_eq!(check_release_exists("v14.0.0".to_string()).unwrap(), true);
-    }
-
-    #[test]
-    fn test_check_release_exists_fail() {
-        assert_eq!(check_release_exists("v14.0.8".to_string()).unwrap(), false);
     }
 }
